updated to use our own logger that defaults to not logging
