+++ death.go
@@ -12,7 +12,7 @@ import (
 	"sync"
 	"time"
 
-	log "github.com/sirupsen/logrus"
+	"github.com/grindlemire/death/deathlog"
 )
 
 // Death manages the death of your application.
@@ -21,15 +21,7 @@ type Death struct {
 	sigChannel  chan os.Signal
 	callChannel chan struct{}
 	timeout     time.Duration
-	log         Logger
-}
-
-// Logger interface to log.
-type Logger interface {
-	Error(v ...interface{})
-	Debug(v ...interface{})
-	Info(v ...interface{})
-	Warn(v ...interface{})
+	log         deathlog.Logger
 }
 
 // closer is a wrapper to the struct we are going to close with metadata
@@ -47,7 +39,7 @@ func NewDeath(signals ...os.Signal) (death *Death) {
 		sigChannel:  make(chan os.Signal, 1),
 		callChannel: make(chan struct{}, 1),
 		wg:          &sync.WaitGroup{},
-		log:         log.StandardLogger()}
+		log:         deathlog.DefaultLogger()}
 	signal.Notify(death.sigChannel, signals...)
 	death.wg.Add(1)
 	go death.listenForSignal()
@@ -61,7 +53,7 @@ func (d *Death) SetTimeout(t time.Duration) *Death {
 }
 
 // SetLogger Overrides the default logger (seelog)
-func (d *Death) SetLogger(l Logger) *Death {
+func (d *Death) SetLogger(l deathlog.Logger) *Death {
 	d.log = l
 	return d
 }
